+++ CHANGELOG.md
@@ -5,7 +5,8 @@ This file is used to list changes made in each version of the PHP cookbook.
 ## Unreleased
 
 - Add support for Ubuntu 20.04
-- Refactor version conditionals to use `Chef::VersionConstraint`
+- Only symlink GMP on Ubuntu 16.04 to aid source builds
+- Drop support for installing from source on Debian 9
 
 ## 7.2.0 (2020-06-19)